requirements.txt
Flask
python-docx
PyPDF2
spacy
fuzzywuzzy
python-Levenshtein
